automation-framework/requirements.txt
pytest>=8.0
pytest-playwright>=0.5.0
pytest-base-url>=2.0
pytest-html>=4.0
httpx>=0.27.0
requests>=2.32.0
pydantic>=2.7.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
allure-pytest>=2.13.5
pytest-xdist>=3.6.1
pytest-rerunfailures>=14.0
